% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{add_samples_db2}
\alias{add_samples_db2}
\title{Add samples from the sample log to the plate with multiplication}
\usage{
add_samples_db2(plate, logIds, dil = c(1, 1), namestyle = 1, group = NA)
}
\arguments{
\item{plate}{PlateObj}

\item{logIds}{A vector of log IDs from the sample log.}

\item{dil}{A vector with length corresponding number of repeats. See details.}

\item{namestyle}{A numeric value indicating the naming style. 1 for long names, 2 for short names.}

\item{group}{A string for bioanalytical group.}
}
\description{
Add samples from the sample log to the plate with multiplication
}
\details{
This function is wrapper around `add_samples_db()` that allows for quick replication of samples by dilution factor vector. 
For instance, it dil = c(1,10), the samples will repeated twice with one fold and 10 fold dilution factor each time.
}
