% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{area_report.PeakRes}
\alias{area_report.PeakRes}
\title{gt table of areas}
\usage{
area_report.PeakRes(
  peaks_res,
  normalize = TRUE,
  blanks = T,
  analytes = T,
  standards = T,
  QCs = T,
  compounds = NULL
)
}
\description{
gt table of areas
}
