% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_residuals.R
\name{calc_var_summary}
\alias{calc_var_summary}
\title{Calculate Summary Statistics for Each Concentration Level For Either Concentration, Area, or Area Ratio}
\usage{
calc_var_summary(
  df,
  col = "conc",
  acc_cutoff = 0.2,
  dev_cutoff = 0.2,
  type = "QC"
)
}
\arguments{
\item{df}{Data frame with columns: stdconc (standardized concentration), conc (concentration), area (peak area), area_ratio (area ratio)}

\item{col}{Column to calculate summary for ("conc", "area", or "area_ratio")}

\item{acc_cutoff}{Accuracy threshold (default is 20\%) for concentration vs standard concentration}

\item{dev_cutoff}{Deviation threshold (default is 20\%) for concentration vs standard concentration}

\item{type}{Type of samples to include ("Standard", "QC", "DQC")}
}
\description{
Calculate Summary Statistics for Each Concentration Level For Either Concentration, Area, or Area Ratio
}
\author{
Omar I. Elashkar
}
