% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_integerate.R
\name{check_chrom_cmpds}
\alias{check_chrom_cmpds}
\title{Check Matching of Compound and Transitions in chrom_res and method database}
\usage{
check_chrom_cmpds(chrom_res, method_id)
}
\arguments{
\item{chrom_res}{ChromRes object}

\item{method_id}{Method ID in the method database
This is important to give no error before merging quantification results to ensure consistency.}
}
\value{
TRUE if all compounds and transitions match, otherwise FALSE
}
\description{
Check Matching of Compound and Transitions in chrom_res and method database
}
