% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{combine_plates}
\alias{combine_plates}
\title{Combine plates in MultiPlate object}
\usage{
combine_plates(plates)
}
\arguments{
\item{plates}{list of PlateObj objects}
}
\value{
MultiPlate object
}
\description{
Combine plates in MultiPlate object
}
