% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_studydesign.R
\name{create_new_study}
\alias{create_new_study}
\title{Create a new study in the database}
\usage{
create_new_study(df)
}
\arguments{
\item{df}{A data frame with one row containing study details:
type, title, description, pkstudy (logical), subject_type}
}
\value{
A data frame with the created study details including generated id, start_date, and status
}
\description{
Create a new study in the database
}
