% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injec_list.R
\name{.add_every_n}
\alias{.add_every_n}
\title{Interject dataframe every Nth position}
\usage{
.add_every_n(df, add_df, every_n)
}
\arguments{
\item{df}{original dataframe}

\item{add_df}{dataframe to add}

\item{every_n}{number of rows to interject add_df}
}
\description{
Interject dataframe every Nth position
}
