% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_integerate.R
\name{.auto_peak}
\alias{.auto_peak}
\title{Calculate Initial Peak}
\usage{
.auto_peak(chrom_res, RT_df = NULL, cutoff = 100)
}
\arguments{
\item{chrom_res}{ChromRes object with single sample and multiple compounds}
}
\description{
Calculate initial peak in chromatogram. This function will use the dataframe with min and max RT values to calculate the initial peak in the chromatogram.
}
\author{
Omar Elashkar
}
