% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{.check_qcs}
\alias{.check_qcs}
\title{Check the quality control samples valid}
\usage{
.check_qcs(std_vec, loq_conc, lqc_conc, mqc_conc, hqc_conc)
}
\arguments{
\item{std_vec}{vector of calibration standards}

\item{loq_conc}{limit of quantification}

\item{lqc_conc}{low quality control concentration}

\item{mqc_conc}{medium quality control concentration}

\item{hqc_conc}{high quality control concentration}
}
\description{
Check the quality control samples valid
}
