% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_parsers.R
\name{.child_to_df}
\alias{.child_to_df}
\title{Utility to extract named xml child to dataframe and rename}
\usage{
.child_to_df(xmltree, tag)
}
\arguments{
\item{xmltree}{xml tree}

\item{tag}{xml tag to extract}
}
\description{
Utility to extract named xml child to dataframe and rename
}
\keyword{internal}
