% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_misc.R
\name{.filter_cmpd}
\alias{.filter_cmpd}
\title{Filter Compounds}
\usage{
.filter_cmpd(peaks_res, cmpd_number)
}
\arguments{
\item{peaks_res}{PeakRes object}

\item{cmpd_number}{numeric vector of compound numbers}
}
\description{
Filter compounds
}
