% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_integerate.R
\name{.filter_peak}
\alias{.filter_peak}
\title{Filter peaks to specific RT range}
\usage{
.filter_peak(
  chrom_res,
  transition_id,
  samples_ids = NULL,
  peak_start,
  peak_end,
  smoothed = FALSE
)
}
\arguments{
\item{chrom_res}{ChromRes object}

\item{transition_id}{Transition ID}

\item{samples_ids}{Sample ID. If NULL, all samples will be used}

\item{peak_start}{Minimum RT value. If NULL, all RT values will be used}

\item{peak_end}{Maximum RT value. If NULL, all RT values will be used}

\item{smoothed}{Logical. If TRUE, use smoothed chromatogram. Default is FALSE}
}
\value{
Dataframe with RT, intensity and sample_id
}
\description{
Filter peaks to specific RT range
}
\keyword{internal}
