% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_integerate.R
\name{.find_max_intensity}
\alias{.find_max_intensity}
\title{Find max intensity for a sample
Observed RT regarless manual or automatic}
\usage{
.find_max_intensity(intensities)
}
\arguments{
\item{intensities}{Dataframe coming from .filter_peak}
}
\description{
Find max intensity for a sample
Observed RT regarless manual or automatic
}
