% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injec_list.R
\name{.injecList}
\alias{.injecList}
\title{injection List Object}
\usage{
.injecList(df, plates)
}
\arguments{
\item{df}{dataframe}

\item{plates}{vector of plate IDs}
}
\description{
injection List Object
}
