% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_integerate.R
\name{.integerate_individual_slack}
\alias{.integerate_individual_slack}
\title{Integerate single sample}
\usage{
.integerate_individual_slack(
  chrom_res,
  compound_id,
  sample_id,
  peak_start,
  peak_end,
  manual = TRUE
)
}
\description{
Integerate single sample
}
