% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_integerate.R
\name{.integerate_next_slack}
\alias{.integerate_next_slack}
\title{Integerate Next Peak}
\usage{
.integerate_next_slack(
  chrom_res,
  compound_id,
  sample_id,
  peak_start,
  peak_end,
  manual = TRUE
)
}
\description{
Integerate Next Peak
}
