% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_file.R
\name{.parse_cmpds}
\alias{.parse_cmpds}
\title{Create a new methods database from a YAML file}
\usage{
.parse_cmpds(path)
}
\arguments{
\item{path}{Path to the YAML file containing method information}
}
\description{
Create a new methods database from a YAML file
}
\keyword{internal}
