% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_parsers.R
\name{.parse_tlynx_csv}
\alias{.parse_tlynx_csv}
\title{Parse targetlynx CSV file}
\usage{
.parse_tlynx_csv(filepath, first_cmpd)
}
\arguments{
\item{filepath}{path to the targetlynx CSV file}

\item{first_cmpd}{name of the first compound, if not provided it will be extracted from the file}
}
\value{
a list of data frames, each containing the data for a compound
}
\description{
Parse targetlynx CSV file
}
\keyword{internal}
