% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_parsers.R
\name{.parse_tlynx_xml}
\alias{.parse_tlynx_xml}
\title{Parse targetlynx}
\usage{
.parse_tlynx_xml(xmlpath, drop_prefix = FALSE)
}
\arguments{
\item{xmlpath}{xml targetlynx output}

\item{drop_prefix}{logical. If TRUE, drop the prefix from the sample name}
}
\value{
A list of Dataframe with each compound
}
\description{
Peaks must be integrated and checked
}
\keyword{internal}
