% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{.plate}
\alias{.plate}
\title{Plate object constructor}
\usage{
.plate(
  m,
  df,
  plate_id,
  empty_rows = NULL,
  last_modified = Sys.time(),
  descr = ""
)
}
\arguments{
\item{m}{a 96-well matrix}

\item{df}{data.frame contains plate's metadata}

\item{empty_rows}{a vector for current active rows}
}
\description{
Plate object constructor
}
