% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_residuals.R
\name{estim_dil_limit}
\alias{estim_dil_limit}
\title{Estimate Dilution Limit Based on Additive and Proportional Errors and LLOQ}
\usage{
estim_dil_limit(add_err, prop_err, lloq)
}
\arguments{
\item{add_err}{Additive error (constant)}

\item{prop_err}{Proportional error (CV)}

\item{lloq}{Lower limit of quantification}
}
\description{
Estimate Dilution Limit Based on Additive and Proportional Errors and LLOQ
}
\examples{
estim_dil_limit(add_err=0.1, prop_err=0.1, lloq=1)
estim_dil_limit(add_err=1, prop_err=0.1, lloq=55)
}
\author{
Omar I. Elashkar
}
