% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{export_integeration}
\alias{export_integeration}
\title{Export Expected RT}
\usage{
export_integeration(chrom_res, path)
}
\arguments{
\item{chrom_res}{ChromRes object}

\item{path}{Path to save the file}
}
\description{
Export expected RT values for each peak in the chromatogram.
}
