% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples_profiles.R
\name{export_pk_profiles}
\alias{export_pk_profiles}
\title{Export PK profiles for a given compound in a specified format
Currently supports "nonmem" format. The exported file will include a CSV with the PK data and an Excel file with the codebook.}
\usage{
export_pk_profiles(x, compound_id, format = "NONMEM", filename = "data.zip")
}
\arguments{
\item{x}{QuantRes object}

\item{compound_id}{Compound ID for which to export PK profiles}

\item{format}{Format to export (currently only "NONMEM" supported)}

\item{filename}{Name of the output zip file (default: "data.zip")}
}
\description{
Export PK profiles for a given compound in a specified format
Currently supports "nonmem" format. The exported file will include a CSV with the PK data and an Excel file with the codebook.
}
\author{
Omar I. Elashkar
}
