% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatogram.R
\name{filter_chrom}
\alias{filter_chrom}
\title{title Filter Chromatogram Peaks}
\usage{
filter_chrom(
  chrom_res,
  transitions_ids = NULL,
  samples_id = NULL,
  cmpd_ids = NULL
)
}
\arguments{
\item{chrom_res}{ChromRes object}

\item{transitions_ids}{Vector of transition IDs to filter. If NULL, all transitions are returned.}

\item{samples_id}{Sample ID to filter.}

\item{cmpd_ids}{Compound ID to filter. It must be numeric. If NULL, all compounds are returned.}
}
\description{
This function filters chromatogram peaks based on transition ID and sample ID.
}
