% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_residuals.R
\name{fit_var}
\alias{fit_var}
\title{Estimate Additive and proportional errors from calibration data}
\usage{
fit_var(
  data,
  level = 0.95,
  method = "nlminb",
  bootstrap = FALSE,
  n_boot = 1000
)
}
\arguments{
\item{data}{Data frame with columns: conc (concentration), stdconc (standardized concentration, e.g. conc/LLOQ)}

\item{level}{Confidence level for the CI (default is 0.95)}

\item{method}{Optimization method (default is "nlminb")}

\item{bootstrap}{Logical indicating whether to perform bootstrap (default is TRUE)}

\item{n_boot}{Number of bootstrap samples (default is 1000)}
}
\description{
Estimate Additive and proportional errors from calibration data
}
\author{
Omar I. Elashkar
}
