% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_residuals.R
\name{formated_print}
\alias{formated_print}
\title{Format and print the results of fit_var}
\usage{
formated_print(x, digits = 3)
}
\arguments{
\item{x}{Data frame with results}

\item{digits}{Number of digits to display}
}
\description{
Format and print the results of fit_var
}
\author{
Omar I. Elashkar
}
