% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_utils.R
\name{get_compound_ID}
\alias{get_compound_ID}
\title{Find Compound ID from compound Name}
\usage{
get_compound_ID(chrom_res, compound_name)
}
\arguments{
\item{chrom_res}{ChromRes object}

\item{compound_name}{Compound Name}
}
\description{
This function returns the compound ID
}
