% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_utils.R
\name{get_sample_ID}
\alias{get_sample_ID}
\title{Find Sample ID from sample Name}
\usage{
get_sample_ID(chrom_res, sample_name)
}
\arguments{
\item{chrom_res}{ChromRes object}

\item{sample_name}{Sample Name}
}
\description{
This function returns the sample ID
}
