% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_utils.R
\name{get_sample_names}
\alias{get_sample_names}
\title{Find sample names for all samples}
\usage{
get_sample_names(chrom_res)
}
\arguments{
\item{chrom_res}{ChromRes object}
}
\value{
data.frame with sample and sample_id
}
\description{
Find sample names for all samples
}
