% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_utils.R
\name{has_default_RT}
\alias{has_default_RT}
\title{check if default expected RT  is set for a compound}
\usage{
has_default_RT(chrom_res, compound_id)
}
\arguments{
\item{chrom_res}{ChromRes object}

\item{compound_id}{Compound ID}
}
\description{
check if default expected RT  is set for a compound
}
