% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{install_py_dep}
\alias{install_py_dep}
\title{Install Python dependencies for PKbioanalysis}
\usage{
install_py_dep(..., envname = "PKbioanalysis")
}
\arguments{
\item{...}{Additional arguments passed to reticulate::py_install}

\item{envname}{Name of the virtual environment to create/use. Default is "PKbioanalysis"}
}
\value{
None
}
\description{
Install Python dependencies for PKbioanalysis
}
