% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatogram.R
\name{is_integrated,ChromRes-method}
\alias{is_integrated,ChromRes-method}
\title{rdname is_integrated}
\usage{
\S4method{is_integrated}{ChromRes}(chrom_res, compound_id, sample_id = NULL)
}
\description{
rdname is_integrated
}
