% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_utils.R
\name{is_integrated}
\alias{is_integrated}
\title{Check if peak has been integrated}
\usage{
is_integrated(chrom_res, compound_id, sample_id = NULL)
}
\arguments{
\item{chrom_res}{ChromRes object}

\item{compound_id}{Compound ID}

\item{sample_id}{Sample ID. If NULL, all samples are checked}
}
\value{
logical
}
\description{
This function checks if the peak has been integrated for a specific compound.
}
\examples{
\dontrun{
lapply(1:10, \(x) is_integrated(chrom_res, sample_id = 1, compound_id = 1))
}
}
