% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_utils.R
\name{is_smoothed}
\alias{is_smoothed}
\title{Return an indicator if the chromatogram is smoothed}
\usage{
is_smoothed(chrom_res)
}
\arguments{
\item{chrom_res}{ChromRes object}
}
\description{
Return an indicator if the chromatogram is smoothed
}
