% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates_class.R
\name{length,MultiPlate-method}
\alias{length,MultiPlate-method}
\title{Length method for MultiPlate}
\usage{
\S4method{length}{MultiPlate}(x)
}
\arguments{
\item{x}{MultiPlate object}
}
\value{
number of plates
}
\description{
Length method for MultiPlate
}
