% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_studydesign.R
\name{make_metabolic_study}
\alias{make_metabolic_study}
\title{Create a metabolic study layout}
\usage{
make_metabolic_study(
  study = "Metabolic Study",
  cmpds,
  time_points = c(0, 5, 10, 15, 30, 45, 60, 75, 90, 120),
  dose = NA,
  n_NAD = 3,
  n_noNAD = 2
)
}
\arguments{
\item{cmpds}{vector of compounds, including any standards}

\item{time_points}{vector of time points}

\item{n_NAD}{number of NAD positive samples. Default is 3}

\item{n_noNAD}{number of NAD negative samples. Default is 2}
}
\value{
MultiPlate object
}
\description{
Create a metabolic study layout
}
\details{
Note that this function does not require plate object. It will create a plate object automatically and return MultiPlate object
}
