% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{multi_plate_study}
\alias{multi_plate_study}
\title{Create a multi-plate study}
\usage{
multi_plate_study(samples)
}
\arguments{
\item{samples}{vector of samples names}
}
\description{
Create a multi-plate study
}
