% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples_profiles.R
\name{nca_table}
\alias{nca_table}
\title{Calculate Cmax, Tmax and AUC for each subject given a compound's PK profiles}
\usage{
nca_table(x, compound_id)
}
\arguments{
\item{x}{QuantRes object with PK profiles extracted}

\item{compound_id}{Compound ID for which to calculate NCA parameters}
}
\value{
data frame with columns: subject_id, cmax, tmax, auc_last, compound_id
}
\description{
Calculate Cmax, Tmax and AUC for each subject given a compound's PK profiles
}
\details{
This function calculates Cmax, Tmax and AUC for each subject given a compound's PK profiles.
}
