% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{percision_per_vial}
\alias{percision_per_vial}
\title{Percision per vial}
\usage{
percision_per_vial(peaks_res, suitability = F)
}
\arguments{
\item{peaks_res}{PeakRes object}

\item{suitability}{logical. If TRUE, suitability samples are ignored}
}
\value{
ggplot2 object
}
\description{
Percision per vial
}
