% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_app.R
\name{plate_app}
\alias{plate_app}
\title{bioanalytic_app}
\usage{
plate_app()
}
\value{
A shiny app. No default return value. Can return a PlateObj if reuse_plate_button is clicked
}
\description{
This function creates a shiny app for plate management
}
