% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{plate_metadata}
\alias{plate_metadata}
\title{Set plate description}
\usage{
plate_metadata(plate, descr)
}
\arguments{
\item{plate}{PlateObj}

\item{descr}{character. Description of the plate}
}
\value{
PlateObj
}
\description{
Set plate description
}
