% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{plate_tree}
\alias{plate_tree}
\title{Plot the design of the plate}
\usage{
plate_tree(plate, plot = TRUE)
}
\arguments{
\item{plate}{PlateObj object}

\item{plot}{logical. If TRUE, plot the tree}
}
\value{
data.tree Node object or DiagrammeR object
plot_tree will focus only on bioanalytical vial types, namely blanks, analytes, standards, QCs. 
The tree order will be plate_id, then group, then vial type, then entity, then number of technical replicates.
}
\description{
Plot the design of the plate
}
