% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{plot_design}
\alias{plot_design}
\title{Plot the design of the plate}
\usage{
plot_design(plate)
}
\arguments{
\item{plate}{PlateObj object}
}
\value{
DiagrammeR object
}
\description{
Plot the design of the plate
}
