% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{plot_peak_areas.PeakRes}
\alias{plot_peak_areas.PeakRes}
\title{Plot peak areas}
\usage{
plot_peak_areas.PeakRes(
  peaks_res,
  normalize = TRUE,
  compounds = NULL,
  blanks = T,
  analytes = T,
  standards = T,
  QCs = T,
  type = "bar",
  ...
)
}
\arguments{
\item{peaks_res}{PeakRes object}

\item{normalize}{logical. If TRUE, normalize the peak area by the IS area.}

\item{blanks}{logical. If TRUE, plot blanks}

\item{analytes}{logical. If TRUE, plot analytes}

\item{standards}{logical. If TRUE, plot standards}

\item{QCs}{logical. If TRUE, plot QCs}

\item{type}{character. Either "bar" or "line"}
}
\value{
ggplot2 object
}
\description{
Plot peak areas
}
