% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_residuals.R
\name{plot_res_rel}
\alias{plot_res_rel}
\title{Plot Relationship Between Concentration and CV/SD}
\usage{
plot_res_rel(df, title)
}
\arguments{
\item{df}{Data frame with columns: stdconc (standardized concentration), cv (coefficient of variation), sdev (standard deviation), Type (e.g., "Estimated", "Observed")}

\item{title}{Plot title}
}
\description{
Plot Relationship Between Concentration and CV/SD
}
\author{
Omar I. Elashkar
}
