% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_anomaly_plots.R
\name{plot_RT.ChromRes}
\alias{plot_RT.ChromRes}
\title{Plotting RT intervals of chromatogram}
\usage{
plot_RT.ChromRes(chrom_res)
}
\arguments{
\item{chrom_res}{ChromRes object}
}
\description{
Plotting RT intervals of chromatogram
}
