% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_misc.R
\name{plot_RT.PeakRes}
\alias{plot_RT.PeakRes}
\title{Plot RT}
\usage{
plot_RT.PeakRes(
  peaks_res,
  normalize = T,
  blanks = T,
  analytes = T,
  standards = T,
  QCs = T,
  facet = F,
  compounds = NULL
)
}
\arguments{
\item{peaks_res}{PeakRes object}
}
\value{
ggplot2 object
}
\description{
Plot RT
}
