% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{precision}
\alias{precision}
\title{Calculate precision (Coefficient of variation)}
\usage{
precision(x, percent = TRUE)
}
\arguments{
\item{x}{vector}

\item{percent}{To return the value as percentage}
}
\value{
numeric
}
\description{
Calculate precision (Coefficient of variation)
}
\details{
A simple calculation of the coefficient of variation (CV) is done
as the standard deviation divided by the mean. By default, the result is in percentage.
}
