% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_object.R
\name{prefilter_precision_data,data.frame-method}
\alias{prefilter_precision_data,data.frame-method}
\title{rdname prefilter_precision_data}
\usage{
\S4method{prefilter_precision_data}{data.frame}(x, type, acc_cutoff = 0.2, dev_cutoff = 0.2)
}
\description{
rdname prefilter_precision_data
}
