% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_object.R
\name{prefilter_precision_data,QuantRes-method}
\alias{prefilter_precision_data,QuantRes-method}
\title{rdname prefilter_precision_data}
\usage{
\S4method{prefilter_precision_data}{QuantRes}(
  x,
  type,
  acc_cutoff = 0.2,
  dev_cutoff = 0.2,
  compound_id = NULL
)
}
\description{
rdname prefilter_precision_data
}
