% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{prefilter_precision_data}
\alias{prefilter_precision_data}
\title{Filter data}
\usage{
prefilter_precision_data(
  x,
  type,
  acc_cutoff = 0.2,
  dev_cutoff = 0.2,
  compound_id = NULL
)
}
\arguments{
\item{x.}{Dataframe or QuantRes Object}

\item{type.}{QC, DQC, or Standard}

\item{cutoff.}{Accuracy cutoff. 20\% by default}
}
\value{
Filtered data
}
\description{
Filter data
}
\author{
Omar I. Elashkar
}
