% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_app-DESKTOP-69BS1LV.R, R/quant_app.R
\name{quant_app}
\alias{quant_app}
\title{quant_app}
\usage{
quant_app()

quant_app()
}
\arguments{
\item{quantres}{A QuantRes object}
}
\description{
This function creates a shiny app for peak integration.

This function creates a shiny app for peak integration.
}
