% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates_class.R
\name{register_plate,MultiPlate-method}
\alias{register_plate,MultiPlate-method}
\title{Register a multiple plates at once}
\usage{
\S4method{register_plate}{MultiPlate}(plate)
}
\arguments{
\item{plate}{MultiPlate object}
}
\value{
a list of RegisteredPlate objects
}
\description{
Register a multiple plates at once
}
\keyword{internal}
