% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates_class.R
\name{register_plate,PlateObj-method}
\alias{register_plate,PlateObj-method}
\title{Register a plate
This will save the plate to the database}
\usage{
\S4method{register_plate}{PlateObj}(plate)
}
\arguments{
\item{plate}{PlateObj object}
}
\value{
Registered PlateObj object
}
\description{
Register a plate
This will save the plate to the database
}
