% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{register_plate}
\alias{register_plate}
\title{This will save the plate to the database}
\usage{
register_plate(plate)
}
\arguments{
\item{plate}{PlateObj object or MultiPlate object}
}
\value{
PlateObj object or list of PlateObj objects
}
\description{
This will save the plate to the database
}
