% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearitycheck.R
\name{response_to_conc}
\alias{response_to_conc}
\title{Convert response to concentration}
\usage{
response_to_conc(quantres, compound_id, response)
}
\arguments{
\item{quantres}{QuantRes object}

\item{compound_id}{character}

\item{response}{numeric. Must match the response type used in linearity. Either abs_response or rel_response}
}
\value{
numeric
}
\description{
Convert response to concentration
}
