% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{reuse_plate}
\alias{reuse_plate}
\title{Reuse and refill a plate with the same ID}
\usage{
reuse_plate(id, extra_fill = 0)
}
\arguments{
\item{id}{numeric. Plate ID}

\item{extra_fill}{numeric. Additional spots to be ignored}
}
\description{
Reuse and refill a plate with the same ID
}
