% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearitycheck.R
\name{reverse_predict}
\alias{reverse_predict}
\title{Reverse predict concentration from response}
\usage{
reverse_predict(fit, newdata, intercept)
}
\arguments{
\item{fit}{lm object}

\item{newdata}{vector or data frame with response values}

\item{intercept}{logical. Whether the model has intercept or not}
}
\value{
numeric. Estimated concentration
}
\description{
Reverse predict concentration from response
}
\author{
Omar I. Elashkar
}
