% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatogram.R
\name{smooth_chrom}
\alias{smooth_chrom}
\title{Smooth Chromatogram Peaks}
\usage{
smooth_chrom(chrom_res, filter = "mean", window = 2, iter = 2)
}
\arguments{
\item{chrom_res}{ChromRes object}

\item{filter}{Filter to use. Options are "mean", "median", "savgol", "gaussian"}

\item{window}{Window size for the filter}

\item{iter}{Number of iterations. If 0, no smoothing is applied.}
}
\description{
This function smooths chromatogram peaks using different algorithms.
}
