% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{str_to_vec}
\alias{str_to_vec}
\title{Convert a comma-separated string to a vector}
\usage{
str_to_vec(x, numeric = FALSE)
}
\arguments{
\item{x}{character. Comma-separated string}

\item{numeric}{logical. If TRUE, convert to numeric vector}
}
\description{
Convert a comma-separated string to a vector
}
