% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{study_chart}
\alias{study_chart}
\title{Plot the design of the plate}
\usage{
study_chart(plate)
}
\arguments{
\item{plate}{PlateObj object}
}
\value{
DiagrammeR object
}
\description{
Plot the design of the plate
}
