% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates_class.R
\name{[[,MultiPlate-method}
\alias{[[,MultiPlate-method}
\title{Subsetting method for MultiPlate}
\usage{
\S4method{[[}{MultiPlate}(x, i, j, ...)
}
\arguments{
\item{x}{MultiPlate object}

\item{i}{index}

\item{j}{index}

\item{...}{additional arguments}
}
\value{
PlateObj object
}
\description{
Subsetting method for MultiPlate
}
