% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{suitability_status}
\alias{suitability_status}
\title{Calculate Percision for samples tagged for suitbalility}
\usage{
suitability_status(peaks_res)
}
\arguments{
\item{peaks_res}{PeakRes object}
}
\value{
ggplot2 object
}
\description{
Percision
}
