% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{add_blank}
\alias{add_blank}
\title{Add blank to the plate
Can be either double blank (DB), CS0IS+ or CS1IS-}
\usage{
add_blank(plate, IS = TRUE, analyte = FALSE)
}
\arguments{
\item{plate}{PlateObj object}

\item{IS}{logical. If TRUE, add IS to the well.}

\item{analyte}{logical. If TRUE, add analyte to the well.}
}
\value{
PlateObj
}
\description{
Add blank to the plate
Can be either double blank (DB), CS0IS+ or CS1IS-
}
