% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{add_cs_curve}
\alias{add_cs_curve}
\title{Add calibration curve to the plate}
\usage{
add_cs_curve(plate, plate_std, rep = 1)
}
\arguments{
\item{plate}{PlateObj}

\item{plate_std}{character}
}
\value{
PlateObj
}
\description{
Add calibration curve to the plate
}
\examples{
plate <- generate_96() |>
 add_cs_curve(c(1, 3, 5, 10, 50, 100, 200))
plot(plate)
}
