% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{add_DB}
\alias{add_DB}
\title{Add double blank (DB) to a plate}
\usage{
add_DB(plate, analytical = FALSE, group = NA)
}
\arguments{
\item{plate}{PlateObj object}

\item{analytical}{logical. If TRUE, the blank is bioanalytical, if FALSE it is analytical.}

\item{group}{A string for bioanalytical group.}
}
\value{
PlateObj
}
\description{
Add double blank (DB) to a plate
}
\examples{
plate <- generate_96() |>
add_DB()
}
