% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{add_DB}
\alias{add_DB}
\title{Add double blank (DB) to a plate}
\usage{
add_DB(plate)
}
\arguments{
\item{plate}{PlateObj object}
}
\value{
PlateObj
}
\description{
Add double blank (DB) to a plate
}
\examples{
plate <- generate_96() |>
add_DB()
}
