% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{add_DQC}
\alias{add_DQC}
\title{Add dilution quality control (DQC) to the plate}
\usage{
add_DQC(plate, conc, fac, rep = 5, group = NA)
}
\arguments{
\item{plate}{PlateObj object}

\item{conc}{numeric. Concentration of the DQC well.}

\item{fac}{numeric. Factor of the DQC well.}

\item{rep}{numeric. Number of replicates. Default is 5.}

\item{group}{A string for bioanalytical group.

The current implementation does not check ULOQ or LLOQ boundaries.}
}
\description{
Add dilution quality control (DQC) to the plate
}
