% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{add_qcs}
\alias{add_qcs}
\title{Add quality control samples to the plate}
\usage{
add_qcs(plate, lqc_conc, mqc_conc, hqc_conc, n_qc = 3, qc_serial = TRUE)
}
\arguments{
\item{plate}{PlateObj object}

\item{lqc_conc}{low quality control concentration}

\item{mqc_conc}{medium quality control concentration}

\item{hqc_conc}{high quality control concentration}

\item{n_qc}{number of QC sets. Default is 3}

\item{qc_serial}{logical. If TRUE, QCs are placed serially}
}
\value{
PlateObj
}
\description{
Add quality control samples to the plate
}
