% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{add_samples}
\alias{add_samples}
\title{Add unknown samples to a plate}
\usage{
add_samples(plate, samples, time = NA, conc = NA, factor = NA, prefix = "S")
}
\arguments{
\item{plate}{PlateObj}

\item{samples}{A vector representing samples names}

\item{time}{A vector representing time points}

\item{conc}{A vector representing concentration}

\item{factor}{A vector representing factor}

\item{prefix}{A prefix to be added before samples names. Default is "S"}
}
\value{
PlateObj
}
\description{
Add unknown samples to a plate
}
\details{
final name will be of form. Prefix-SampleName-Time-Concentration-Factor
}
\examples{
plate <- generate_96() |>
 add_samples(paste0("T", 1:12))
}
