% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{add_suitability}
\alias{add_suitability}
\title{Add suitability sample to the plate}
\usage{
add_suitability(plate, conc, label = "suitability")
}
\arguments{
\item{plate}{PlateObj object.}

\item{conc}{numeric. Concentration of the suitability well.}

\item{label}{character. Label for the suitability well. Default is "suitability".}
}
\value{
PlateObj
}
\description{
Add suitability sample to the plate
}
