% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_misc.R
\name{area_report.PeakRes}
\alias{area_report.PeakRes}
\title{gt table of areas}
\usage{
area_report.PeakRes(
  peaks_res,
  normalize = TRUE,
  blanks = T,
  analytes = T,
  standards = T,
  QCs = T,
  compounds = NULL
)
}
\arguments{
\item{peaks_res}{PeakRes object}

\item{normalize}{logical. If TRUE, normalize the peak area by the IS area.}

\item{blanks}{logical. If TRUE, include blanks}

\item{analytes}{logical. If TRUE, include analytes}

\item{standards}{logical. If TRUE, include standards}

\item{QCs}{logical. If TRUE, include QCs}

\item{compounds}{numeric vector of compound numbers to include. If NULL, include all compounds}
}
\description{
gt table of areas
}
