% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{clear_python_deps}
\alias{clear_python_deps}
\title{Clear Managed Python Environments}
\usage{
clear_python_deps()
}
\description{
Removes the ephemeral Python environments and cached artifacts managed 
by reticulate and uv.
}
