% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injec_list.R
\name{download_sample_list}
\alias{download_sample_list}
\title{Download sample list from database to local spreadsheet}
\usage{
download_sample_list(sample_list, vendor)
}
\arguments{
\item{sample_list}{dataframe of sample list either from db or from write_injec_seq}

\item{vendor}{currently only 'masslynx' and 'masshunter' are supported}
}
\value{
dataframe
}
\description{
Download sample list from database to local spreadsheet
}
\details{
For 'masslynx' and 'masshunter', the exported format will be in csv format, compatible with the respective software.
}
