% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_misc.R
\name{export_run.PeakRes}
\alias{export_run.PeakRes}
\title{Export run}
\usage{
export_run.PeakRes(peaks_res, path)
}
\arguments{
\item{peaks_res}{PeakRes object}

\item{path}{path to save csv}
}
\description{
Export run
}
