% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_integerate.R
\name{extract_peak_bounds}
\alias{extract_peak_bounds}
\title{Extract Peak Boundaries}
\usage{
extract_peak_bounds(chrom_res, compound_id)
}
\arguments{
\item{chrom_res}{ChromRes object}

\item{compound_id}{Compound ID
The function automatically priortizes observed peak boundaries (manual integration) over expected ones.
If observed boundaries are not available, it falls back to expected boundaries.}
}
\value{
Dataframe with compound_id, min, max
}
\description{
Extract peak boundaries for a given compound ID
}
