% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{generate_96}
\alias{generate_96}
\title{Generate 96 Plate
Generate a typical 96 well plate. User need to specify the empty rows which a going to be used across the experiment.}
\usage{
generate_96(descr = "", start_row = "A", start_col = 1)
}
\arguments{
\item{descr}{plate description.}

\item{start_row}{A letter corresponding to empty rows in a 96 well plate. Default is A.}

\item{start_col}{A number indicating a column number to start with, given the start row. Default is 1.}
}
\value{
PlateObj
}
\description{
Generate 96 Plate
Generate a typical 96 well plate. User need to specify the empty rows which a going to be used across the experiment.
}
\examples{
plate <- generate_96()
plot(plate)

plate <- generate_96("calibration", start_row = "C", start_col = 11)
plot(plate)

}
