% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{generate_96}
\alias{generate_96}
\title{Generate 96 Plate
Generate a typical 96 well plate. User need to specify the empty rows which a going to be used across the experiment.}
\usage{
generate_96(descr = "", empty_rows = NULL, extra_fill = 0)
}
\arguments{
\item{descr}{plate description.}

\item{empty_rows}{vector of letters corresponding to empty rows in a 96 well plate.}

\item{extra_fill}{additional spots to be ignored from the first empty row.}
}
\value{
PlateObj
}
\description{
Generate 96 Plate
Generate a typical 96 well plate. User need to specify the empty rows which a going to be used across the experiment.
}
\examples{
plate <- generate_96()
plot(plate)

plate <- generate_96("calibration", empty_rows = c("C", "D", "E"), extra_fill = 11)
plot(plate)

}
