% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_integerate.R
\name{integrate}
\alias{integrate}
\title{Integerate Peak with trapzoid method given start and end}
\usage{
integrate(chrom_res, compound_id, samples_ids, smoothed = TRUE)
}
\arguments{
\item{chrom_res}{ChromRes object. Must have observed RT values}

\item{compound_id}{Compound ID}

\item{samples_ids}{Sample ID. If NULL, all samples will be used}

\item{smoothed}{Logical. If TRUE, use smoothed chromatogram. Default is TRUE}
}
\description{
Integerate Peak with trapzoid method given start and end
}
