% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKChromaMetrics-package.R,
%   R/PKbioanalysis-package.R
\docType{package}
\name{PKbioanalysis-package}
\alias{PKbioanalysis-package}
\title{PKbioanalysis: Pharmacokinetic Bioanalysis Experiments Design and Exploration}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Automate pharmacokinetic/pharmacodynamic bioanalytical procedures based on best practices and regulatory recommendations. The package impose regulatory constrains and sanity checking for common bioanalytical procedures. Additionally, 'PKbioanalysis' provides a relational infrastructure for plate management and injection sequence.

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Automate pharmacokinetic/pharmacodynamic bioanalytical procedures based on best practices and regulatory recommendations. The package impose regulatory constrains and sanity checking for common bioanalytical procedures. Additionally, 'PKbioanalysis' provides a relational infrastructure for plate management and injection sequence.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://omarashkar.github.io/PKbioanalysis/}
  \item Report bugs at \url{https://github.com/OmarAshkar/PKbioanalysis/issues}
}


Useful links:
\itemize{
  \item \url{https://omarashkar.github.io/PKbioanalysis/}
  \item Report bugs at \url{https://github.com/OmarAshkar/PKbioanalysis/issues}
}

}
\author{
\strong{Maintainer}: Omar Elashkar \email{omar.i.elashkar@gmail.com} (\href{https://orcid.org/0000-0002-5505-778X}{ORCID})

}
\keyword{internal}
