% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_misc.R
\name{plot_RT.PeakRes}
\alias{plot_RT.PeakRes}
\title{Plot RT}
\usage{
plot_RT.PeakRes(
  peaks_res,
  normalize = TRUE,
  blanks = TRUE,
  analytes = TRUE,
  standards = TRUE,
  QCs = TRUE,
  facet = FALSE,
  compounds = NULL
)
}
\arguments{
\item{peaks_res}{PeakRes object}

\item{normalize}{logical. If TRUE, normalize the peak area by the IS area.}

\item{blanks}{logical. If TRUE, plot blanks}

\item{analytes}{logical. If TRUE, plot analytes}

\item{standards}{logical. If TRUE, plot standards}

\item{QCs}{logical. If TRUE, plot QCs}

\item{facet}{logical. If TRUE, facet by compound name}

\item{compounds}{numeric vector of compound numbers to include. If NULL, include all compounds}
}
\value{
ggplot2 object
}
\description{
Plot RT
}
