% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_misc.R
\name{precision_per_vial}
\alias{precision_per_vial}
\title{Precision per vial}
\usage{
precision_per_vial(peaks_res, suitability = FALSE)
}
\arguments{
\item{peaks_res}{PeakRes object}

\item{suitability}{logical. If TRUE, suitability samples are ignored}
}
\value{
ggplot2 object
}
\description{
Precision per vial
}
