% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/quant_object.R
\name{prefilter_precision_data}
\alias{prefilter_precision_data}
\alias{prefilter_precision_data,QuantRes-method}
\alias{prefilter_precision_data,data.frame-method}
\title{Filter data}
\usage{
prefilter_precision_data(
  x,
  type,
  acc_cutoff = 0.2,
  dev_cutoff = 0.2,
  compound_id = NULL
)

\S4method{prefilter_precision_data}{QuantRes}(
  x,
  type,
  acc_cutoff = 0.2,
  dev_cutoff = 0.2,
  compound_id = NULL
)

\S4method{prefilter_precision_data}{data.frame}(x, type, acc_cutoff = 0.2, dev_cutoff = 0.2)
}
\arguments{
\item{x}{Dataframe or QuantRes Object}

\item{type}{QC, DQC, or Standard}

\item{acc_cutoff}{Accuracy cutoff. 20\% by default}

\item{dev_cutoff}{Deviation cutoff. 20\% by default}

\item{compound_id}{Compound ID to filter. If NULL, all compounds are considered}
}
\value{
Filtered data
}
\description{
Filter data
}
\author{
Omar I. Elashkar
}
