% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_object.R
\name{prefilter_precision_data}
\alias{prefilter_precision_data}
\title{Filter data}
\usage{
prefilter_precision_data(x, type, cutoff = 0.2, compound_id)
}
\arguments{
\item{x.}{Dataframe or QuantRes Object}

\item{type.}{QC, DQC, or Standard}

\item{cutoff.}{Accuracy cutoff. 20% by default}
}
\description{
Filter data
}
