% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_app.R
\name{quant_app}
\alias{quant_app}
\title{Quantification App}
\usage{
quant_app()
}
\arguments{
\item{quantres}{A QuantRes object}
}
\description{
This function creates a shiny app for quantification after peak integeration
}
