% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_parsers.R
\name{read_experiment_results}
\alias{read_experiment_results}
\title{Read experiment results}
\usage{
read_experiment_results(
  x,
  drop_prefix = FALSE,
  vendor = "targetlynx_xml",
  logkey = "Index"
)
}
\arguments{
\item{x}{path to experiment results. See details}

\item{drop_prefix}{logical. If TRUE, drop the prefix from the sample name}

\item{vendor}{vendor name. Currently only "targetlynx_xml" or "targetlynx_csv" are supported.}

\item{logkey}{character. The column name in the targetlynx CSV file that contains the injection sequence or log key. Default is "Index" which is the default column in targetlynx CSV exports, but it can be customized if the user has a different column name for the injection sequence.}
}
\value{
QuantRes object with the results of the experiment.
}
\description{
Read experiment results
}
\details{
Currently only targetlynx XML or CSV exported files are supported.
}
