% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_parsers.R
\name{read_experiment_results}
\alias{read_experiment_results}
\title{Read experiment results}
\usage{
read_experiment_results(x, drop_prefix = FALSE, vendor = "targetlynx_xml")
}
\arguments{
\item{x}{path to experiment results. See details}

\item{drop_prefix}{logical. If TRUE, drop the prefix from the sample name}

\item{vendor}{vendor name. Only "targetlynx"}
}
\value{
QuantRes object with the results of the experiment.
}
\description{
Read experiment results
}
\details{
Currently only targetlynx XML or CSV exported files are supported.
}
