% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_misc.R
\name{run_summary.PeakRes}
\alias{run_summary.PeakRes}
\title{Get Summary of PeakRes object}
\usage{
\method{run_summary}{PeakRes}(object)
}
\arguments{
\item{peaks_res}{PeakRes object}
}
\description{
Get Summary of PeakRes object
}
