% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injec_list.R
\name{write_injec_seq}
\alias{write_injec_seq}
\title{Export injection sequence to vendor specific format}
\usage{
write_injec_seq(injec_seq)
}
\arguments{
\item{injec_seq}{InjecListObj object}
}
\value{
dataframe
}
\description{
Export injection sequence to vendor specific format
}
